requirements.txt
asgiref==3.10.0
beautifulsoup4==4.14.2
bs4==0.0.2
click==8.3.0
colorama==0.4.6
cssbeautifier==1.15.4
Django==5.2.7
django-extensions==4.1
django-nested-admin==4.1.6
django-tinymce==5.0.0
djlint==1.36.4
EditorConfig==0.17.1
jsbeautifier==1.15.4
json5==0.12.1
MarkupSafe==3.0.3
pathspec==0.12.1
python-monkey-business==1.1.0
PyYAML==6.0.3
regex==2025.10.23
ruff==0.14.2
six==1.17.0
soupsieve==2.8
sqlparse==0.5.3
tqdm==4.67.1
typing_extensions==4.15.0
Werkzeug==3.1.3
